% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_hydrus.R
\name{call.H1D}
\alias{call.H1D}
\title{Executes H1D_CALC.EXE from  Hydrus 1D directory}
\usage{
call.H1D(project.path, hydrus.path, ...)
}
\arguments{
\item{...}{}
}
\description{
Executes H1D_CALC.EXE from  Hydrus 1D directory
}

