% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_obs_nodes_out.R
\name{compile.obs_nodes.out}
\alias{compile.obs_nodes.out}
\title{Compile outputs at observation nodes, from all simulations}
\usage{
compile.obs_nodes.out(outputs.path, num.of.sims, output = c("h", "theta"),
  obs.nodes, ...)
}
\arguments{
\item{...}{}
}
\description{
Compile outputs at observation nodes, from all simulations
}

