% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1dproj-class.R
\name{convert_units}
\alias{convert_units}
\title{Convert Units}
\usage{
convert_units(x, from, to)
}
\arguments{
\item{x}{x}

\item{from}{from}

\item{to}{to}
}
\value{
new units
}
\description{
Convert Units
}
