% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bc.R
\name{create.bc}
\alias{create.bc}
\title{Create inputs to set boundary conditions (atmospheric, top, and bottom)}
\usage{
create.bc(
  project.path,
  atmos = TRUE,
  atmos.df = NULL,
  hCritS = 0,
  top.bc.type = NULL,
  top.bc.value = NULL,
  freeD = TRUE,
  bot.bc.type = NULL,
  bot.bc.value = NULL,
  hCritA = NULL,
  ...
)
}
\arguments{
\item{project.path}{Hydrus project path. Must have SELECTOR.IN already, which is made by create.H1D.project().}

\item{atmos}{logical, length = 1. Does this project have atmospheric conditions as a boundary condition? If TRUE, this overrides any argument given
to 'top.bc.type' and 'top.bc.value'. If FALSE, an argument must be supplied to 'top.bc.type'}

\item{atmos.df}{dataframe woith columns for 'time' and one or more columns of 'Prec' (precipitation rate, L/T), 'rSoil' (potential evaporation rate,
L/T), 'rRoot' (potential transpiration rate, L/T). a column 'PET' (potential evapotranspiration rate, L/T), can also be given instead of rSoil
and rRoot, and PET will be split according to 'trans.pc'. If PET is given, rSoil and rRoot will be overwritten. Note that these are all rates (L/T)
not depths per the given time period. \cr
The times in the dataframe are used as the end points of the period (not starting points). As such, the upper row of the dataframe should not
begin at time=0, the first row will define the conditions from time=0 until time[1], and the second row will define conditions from time[1] to
time[2], and so on. If a constant atmospheric flux is desired, then the dataframe will have only one row, with time = max(time), the endpoint of
the simulation, and the values given will be applied for the whole duration.}

\item{hCritS}{numeric, length = 1. the maximum allowed surface head. in project length units. Only required if atmos = TRUE. Default = 0, meaning
that water isn't allowed to pond on the surface. Set this value > 0 if ponding is to be allowed. Do not set this < 0. In the Hydrus GUI 'Water
Flow Boundary Conditions' menu, setting this to 0 is equivalent to selecting 'Atmospheric BC with Surface Run Off' and setting it > 0 is
'Atmospheric BC with Surface Layer'.}

\item{top.bc.type}{character, length = 1. if atmos = TRUE, this is ignored. if atmos = FALSE, then one of four types is expected here: 'ch' constant
head, 'cf' constant flux, 'vh' variable head, 'vf' variable flux. If 'cf' or 'ch' then a numeric value must be supplied to 'top.bc.value'. If 'vf'
or 'vh', then a dataframe must be given to 'top.bc.value'.}

\item{top.bc.value}{either a numeric, length = 1, if top.bc.type = 'cf' or 'ch'; or a dataframe with two columns 'time' and 'head' (vh) or 'flux'
(vf) if top.bc.type = 'vf' or 'vh'. all must obey project units, with flux given as a rate (length/time).}

\item{freeD}{logical, length = 1. default TRUE. is the bottom boundary free draining? meaning should a unity gradient be assigned throughout the simulation?
(i.e., matric potential is at equilibrium, and flux driven by gravity only)0. if this is TRUE, then bot.bc.type is ignored. if this is FALSE, then
bot.bc.type and bot.bc.value must be given.}

\item{bot.bc.type}{character, length = 1. if freeD = TRUE, this is ignored. if freeD = FALSE, then one of four types is expected here: 'ch' constant
head, 'cf' constant flux, 'vh' variable head, 'vf' variable flux. If 'cf' or 'ch' then a numeric value must be supplied to 'bot.bc.value'. If 'vf'
or 'vh', then a dataframe must be given to 'bot.bc.value'.}

\item{hCritA}{numeric, length = 1. the minimum allowed surface head, i.e. a dryness limit. Once this is hit, evaporation will cease. If left NULL,
the default will be set at 1e5 cm, which will be converted to the unit given by SELECTOR.IN's LUnit. This can be left as NULL for nearly all uses.}

\item{trans.pc}{numeric, length = 1, between 0 and 1. If PET is a column in atmos.df, what percent of PET should be assigned to transpiration
(rRoot)? The remainder will be assigned to evaporation (rSoil). Default = 1, i.e. all transpiration. Not required if 'PET' column is not given in
atmos.df.}
}
\description{
This functions houses write.bottom.bc and write.atmosph.in, which work together to set the boundary conditions for a Hydrus simulation. Based on
the supplied arguments, it creates a dataframe of variable boundary conditions, including any atmospheric conditions, sets up the required
blocks of SELECTOR.IN, and writes ATMOSPH.IN. \cr
This is one of three parts to prepping a HYDRUS simulation, with create.H1D project and create.soil.profile. This function must be preceded by
create.H1D.project() because a pre-existing SELECTOR.IN file is required to pull data from, such as LUnit (length unit), TUnit (time unit),
start time (tInit), and end time (tMax), to ensure that the boundary condition data is in alignment. \cr
There are several Hydrus boundary condition settings that are not currently coded in this function, including bottom deep drainage, bottom seepage
face, and bottom flow to drains.
}
