% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_H1D_project.R
\name{create.H1D.project}
\alias{create.H1D.project}
\title{Creates a new hydrus project folder with "HYDRUS1D.DAT", "SELECTOR.IN" and "DESCRIPT.TXT" files}
\usage{
create.H1D.project(
  project.name,
  parent.dir,
  description = NULL,
  TimeUnit = "days",
  SpaceUnit = "cm",
  PrintTimes = 1,
  processes = c(WaterFlow = T, RootWaterUptake = F),
  geometry,
  soil.para,
  soil.model,
  soil.hys,
  initial.cond,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{project.name}{Name of the project}

\item{parent.dir}{Path to the project folder, where a folder with project.name will be saved}

\item{TimeUnit}{Simulation time unit information (default = days). permitted: seconds, minutes, hours, days, years.}

\item{SpaceUnit}{Vertical spatial unit (decault = cm). permitted: mm, cm, m}

\item{PrintTimes}{Time levels at which the outputs should be printed. These are the times that willbe output in the Nod_Inf.out file among others.}

\item{processes}{Main processes (e.g., WaterFlow, SoluteTransport etc) to be simulated}

\item{geometry}{Profile geometry info (Depth, # of nodes and # of obs nodes)}

\item{soil.para}{List of soil hydraulic parameters, one entry per material to be included in simulation. See ?hydrusR::write.hydraulic.para for details.}

\item{soil.model}{integer, 0-7. identifies the soil model to be used. 0 = standard van Genuchten-Mualem. See ?hydrusR::write.hydraulic.para for details.}

\item{soil.hys}{include hysteresis in simulation? integer. 0 = No, 1 = Yes.}

\item{overwrite}{if TRUE, function will not ask to overwrite the project if it already exists. careful.}
}
\description{
Creates a new hydrus project folder with "HYDRUS1D.DAT", "SELECTOR.IN" and "DESCRIPT.TXT" files
}
\examples{

create.H1D.project(project.name = "testproj2", parent.dir = parent_dir, discription = NULL,
                   SpaceUnit = "cm", TimeUnit = "days", PrintTimes = 1,
                   processes = c(WaterFlow = T, SoluteTransport = F, RootWaterUptake = F,
                   RootGrowth = F, Unsatchem = F, HP1 = F, EquillibriumAdsorption = F),
                   initial.cond = c(NumberOfSolutes = 0, InitialCondition = 0),
                   geometry = c(ProfileDepth = 200, NumberOfNodes = 4,
                                ObservationNodes = 0, MaterialNumber = 1, SubregionNumber = 1))
}
\author{
Subodh Acharya <https://github.com/shoebodh>; Trevor Baker <tbaker@iegconsulting.com>
}
