% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1dproj-class.R
\name{create.profile_dat}
\alias{create.profile_dat}
\title{Create PROFILE.DAT}
\usage{
create.profile_dat(H1Dproj, obs.nodes = NULL, Conc = 0)
}
\arguments{
\item{H1Dproj}{H1Dproj}

\item{obs.nodes}{obs.nodes (default: NULL)}

\item{Conc}{concentation (default: 0.0)}
}
\value{
create PROFILE.DAT
}
\description{
Create PROFILE.DAT
}
