% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir.copy}
\alias{dir.copy}
\title{Copy contents of a folder to another}
\usage{
dir.copy(src.dir, dest.dir, overwrite = TRUE, ask = FALSE)
}
\arguments{
\item{src.dir}{Source directory}

\item{dest.dir}{Destination directory}

\item{overwrite}{Logical, if true overwrites the existing directory, default = T}

\item{ask}{Logical, should ask whether to create/delete directories if required, default = FALSE}
}
\value{
Copy contents of a folder to another
}
\description{
Copy contents of a folder to another
}
