% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format2sci}
\alias{format2sci}
\title{Format number to scientific format according to HYDRUS inputs}
\usage{
format2sci(x, ndec, power.digits, ...)
}
\arguments{
\item{x}{A numeric vector to be formatted}

\item{ndec}{Number of decimal places to apply (e.g, .000xx)}

\item{power.digits}{Number of power digits to apply (e.g., e+002)}
}
\description{
Format number to scientific format according to HYDRUS inputs
}
