% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.decimalplaces}
\alias{get.decimalplaces}
\title{get.decimalplaces}
\usage{
get.decimalplaces(x)
}
\arguments{
\item{x}{A double numerical}
}
\value{

}
\description{
Get the number of decimal digits in a numeric value
}
\examples{
get.decimalplaces(10.343434)
}
