% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wtd.R
\name{get.water.table}
\alias{get.water.table}
\title{Extracts water table depth at each time step}
\usage{
get.water.table(tlevel.out, soil.depth)
}
\arguments{
\item{soil.depth}{}
}
\description{
Extracts water table depth at each time step
}

