% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_table.R
\name{get_water_table}
\alias{get_water_table}
\title{Calculate WT depth from soil moisture profile and parameters}
\usage{
get_water_table(simulation.path, sim.num)
}
\arguments{
\item{sim.num}{}
}
\description{
Calculate WT depth from soil moisture profile and parameters
}

