% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1dproj-class.R
\name{h1dproj_initialize}
\alias{h1dproj_initialize}
\title{Constructor method for the h1dproj}
\usage{
\S4method{initialize}{h1dproj}(.Object)
}
\arguments{
\item{.Object}{.Object}
}
\description{
Constructor method for the h1dproj
}
