% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1dproj-class.R
\name{h1dproj_show}
\alias{h1dproj_show}
\title{Show Method for the h1dproj class}
\usage{
\S4method{show}{h1dproj}(object)
}
\arguments{
\item{object}{object}
}
\description{
Show Method for the h1dproj class
}
\examples{
h1dproj()
methods::new("h1dproj")

}
