% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{len}
\alias{len}
\title{aka length()}
\usage{
len(x)
}
\arguments{
\item{x}{any vector class, any length}
}
\value{
integer, length of vector
}
\description{
aka length()
}
\examples{
x <- c(10,1,1,1,2,3,4,5,5,5,5,5)
len(x) # 12
identical(length(x), len(x)) #TRUE
}
