% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg.R
\name{load_pkg}
\alias{load_pkg}
\title{wrapper for pkgload::load_all}
\usage{
load_pkg(dir = "C:/Users/t/Documents/hydrusR")
}
\arguments{
\item{dir}{package directory}
}
\description{
use this when writing new code rather than 'Build/Install and restart'. much faster
}
