% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lu}
\alias{lu}
\title{aka length(unique())}
\usage{
lu(x)
}
\arguments{
\item{x}{any vector class, any length}
}
\value{
integer, how many unique values in the input
}
\description{
aka length(unique())
}
\examples{
x <- c(1,1,1,2,3,4,5,5,5,5,5)
lu(x) # 5
identical(length(unique(x)), lu(x)) #TRUE, both are 5
}
