% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_hourly.R
\name{pet.hourly}
\alias{pet.hourly}
\title{Generates hourly PET inputs from provided daily value.}
\usage{
pet.hourly(pet.daily, hours = 6:20)
}
\arguments{
\item{pet.daily}{numeric vector of daily PET values in same length units as project. i.e., must convert mm PET to cm if your project is in cm.}

\item{hours}{hours of day that will have non-zero ET. numeric vector, not a range. i.e., 6:20 not c(6,20)}
}
\description{
Generates hourly PET inputs from provided daily value.
}
\examples{
pet.val <- pet.hourly(pet.daily = c(1,2,4), hours = 6:20)
plot(pet.val) #show the daily totoals broken into hourly with diurnal variation
sum(pet.val$pet[1:24]) #sum = 1 on day 1
sum(pet.val$pet[25:48]) #sum = 2 on day 2
sum(pet.val$pet[49:72]) #sum = 4 on day 3
}
