% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_moisture.profile}
\alias{plot_moisture.profile}
\title{Plot moisture profile}
\usage{
plot_moisture.profile(project.path, output = "Moisture")
}
\arguments{
\item{project.path}{project.path}

\item{output}{default: "Moisture"}
}
\value{
plot moisture profile
}
\description{
Plot moisture profile
}
