% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_pressure.profile}
\alias{plot_pressure.profile}
\title{Plot pressure profile}
\usage{
plot_pressure.profile(project.path, output = "Head")
}
\arguments{
\item{project.path}{project.path}

\item{output}{default: "Head"}
}
\value{
plot pressure profile
}
\description{
Plot pressure profile
}
