% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile.last}
\alias{plot_profile.last}
\title{Plot profile generic}
\usage{
plot_profile.last(project.paths, output = "Head", measured)
}
\arguments{
\item{output}{default: "Head"}

\item{project.path}{project.path}
}
\value{
ggplot object of profile
}
\description{
Plot profile generic
}
