% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevel_out.R
\name{read.A_level.out}
\alias{read.A_level.out}
\title{Read outputs of ALEVEL.OUT}
\usage{
read.A_level.out(simulation.path, output = NULL, warn = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Read outputs of ALEVEL.OUT
}

