% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obs_node.R
\name{read.obs_node}
\alias{read.obs_node}
\title{Read Outputs of OBS_NOD.OUT}
\usage{
read.obs_node(simulation.path, obs.output = c("h", "theta"), obs.nodes, ...)
}
\arguments{
\item{...}{}
}
\description{
Read Outputs of OBS_NOD.OUT
}

