% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tlevel_out.R
\name{read.tlevel.out}
\alias{read.tlevel.out}
\title{READ outputs of TLEVEL.OUT}
\usage{
read.tlevel.out(
  project.path,
  out.file = "T_Level.out",
  output = NULL,
  warn = FALSE
)
}
\arguments{
\item{project.path}{project.path}

\item{out.file}{out.file (default: "T_Level.out")}

\item{output}{output (default: NULL)}

\item{warn}{default: FALSE}
}
\value{
read "T_Level.out" as data.table
}
\description{
READ outputs of TLEVEL.OUT
}
