% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_h1d.R
\name{run.H1D.example}
\alias{run.H1D.example}
\title{Runs a test hydrus project}
\usage{
run.H1D.example(profile.depth = 200, dz = 1, ntime = 1000, rdepth = 100,
  soil.para = list(thr = 0.045, ths = 0.43, Alfa = 0.145, n = 2.69, Ks = 29.7,
  l = 0.45),
  hydrus.path = "C:/Program Files (x86)/PC-Progress/Hydrus-1D 4.xx", ...)
}
\arguments{
\item{...}{}
}
\description{
Runs a test hydrus project
}
\examples{

run.H1D.example()
}
