% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_soilparam.R
\name{sensitivity_soilparam}
\alias{sensitivity_soilparam}
\title{Sensitivity Analysis of soil-hydraulic parameters}
\usage{
sensitivity_soilparam(project.path, obs_nodes_all, para, i, results, ...)
}
\arguments{
\item{project.path}{folder where the model input files for the current run are stored}

\item{obs_nodes_all}{depths of the observation nodes}

\item{para}{list of soil-hydraulic parameters to be used in the current run}

\item{i}{line in data.frame "results"}

\item{results}{data.frame to store the values of the soil-hydraulic parameters and the resulting RMSE and CEFF}

\item{run}{name of the current run, e.g. "Alfa max"}
}
\value{
2 figures: "Obs_points_'run'" shows timeseries of predicted soil moisture at observation nodes, "Mod_eval_run_'run'" plots observed and predicted soil moisture, with RMSE and CEFF, results
}
\description{
This function conducts a sensitivity analysis, comparing the effect of a max and min value of soil hydraulic input parameters
}
