% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{su}
\alias{su}
\title{aka sort(unique())}
\usage{
su(x)
}
\arguments{
\item{x}{any vector class, any length}
}
\value{
vector of unique values sorted with sort()
}
\description{
aka sort(unique())
}
\examples{
x <- c(10,1,1,1,2,3,4,5,5,5,5,5)
su(x) # 1  2  3  4  5 10
identical(sort(unique(x)), su(x)) #TRUE
}
