% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{u}
\alias{u}
\title{aka unique()}
\usage{
u(x)
}
\arguments{
\item{x}{any vector class, any length}
}
\value{
vector of unique values
}
\description{
aka unique()
}
\examples{
x <- c(1,1,1,2,3,4,5,5,5,5,5)
u(x) #vector 1:5
identical(unique(x), u(x)) #TRUE, both are 1:5
}
