% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unit_mult}
\alias{unit_mult}
\title{Get unit multiples for converting cm and days into project units}
\usage{
unit_mult(SpaceUnit = "cm", TimeUnit = "hours")
}
\value{
vector, length 3. L.multiple, T.multiple, LT.multiple. Multiply cm, d, or cm/d values by these to get L, T, or L/T in your project units.
}
\description{
Get unit multiples for converting cm and days into project units
}
