% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_atmosph_in.R
\name{write.atmosph.in}
\alias{write.atmosph.in}
\title{Write atmospheric boundary condition inputs}
\usage{
write.atmosph.in(project.path, atm.bc.data, hCritS = 0, ...)
}
\arguments{
\item{project.path}{Hydrus project path.}

\item{atm.bc.data}{dataframe containing all time-variable atmosphere records. must have column 'tAtm', which is the time of each record. Full list
of columns is given in Hydrus manual Table 12.9. Currently accepted columns are listed next. Not all of these must be given for every simulation,
only those that are Prec (precipitation rate, L/T), rSoil (potential evaporation rate,
L/T), rRoot (potential transpiration rate, L/T), hCritA (minimum allowed pressure head at soil surface, dryness limit, L), rB (bottom flux rate,
L/T), hB (pressure head at soil bottom, L), hT (pressure head at soil top, L).}

\item{hCritS}{numeric, length = 1. the maximum allowed surface head. in project length units. Default = 0, meaning that watrer isn't allowed to
pond on the surface. Set this value > 0 if ponding is to be allowed. Do not set this < 0.}
}
\description{
This writes ATMOSPH.IN which contains time-variable boundary conditions, including precipitation, potential evaporation and transpiration, boundary
heads and boundary fluxes. This function does not need to be run for all simulations, but most need it. It is not needed if the upper boundary
condition is either a constant head or constant flux (set in write.bottom.bc) and the lower boundary condition is either a constant head or
constant flux (set in write.bottom.bc) or is free draining (set in write.sim.settings). If the simulation has precipitation or evapotranspiration,
then this needs to be run.
}
