% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bc_bottom.R
\name{write.bc.bottom}
\alias{write.bc.bottom}
\title{Writes codes for bottom boundary conditions to SELECTOR.IN}
\usage{
write.bc.bottom(project.path, freeD = freeD, constant.bc, bc.type, bc.value)
}
\arguments{
\item{project.path}{path to the Hydrus project. where SELECTOR.IN file is saved.}

\item{freeD}{logical. is this a free-draining profile. i.e. untiy gradient at bottom.}

\item{constant.bc}{logical, length = 1. is this a time-variable (FALSE) or constant boundary (TRUE) condition? If this is a variable condition, then
additional data will need to be prepared for ATMOSPH.IN by write.atmopsh.in(). If this is a constant boundary condition, then no ATMOSPH.IN data needs
to be set.}

\item{bc.type}{character, length = 1. "flux" or "head". what type of boundary condition is being set?}

\item{bc.value}{numeric, length = 1. the value to be set for a constant flux bottom condition, in project length/time units. Does not need to be
specified for constant head conditions because the Head value of the bottom Node in PROFILE.DAT will be used.}
}
\description{
Writes codes for bottom boundary conditions to SELECTOR.IN
}
\author{
Subodh Acharya <https://github.com/shoebodh>; Trevor Baker <tbaker@iegconsulting.com>
}
