% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_obs_nodes.R
\name{write.obs.nodes}
\alias{write.obs.nodes}
\title{Write observation nodes to profile.dat}
\usage{
write.obs.nodes(project.path, obs.nodes = NULL, ...)
}
\arguments{
\item{project.path}{path to the Hydrus project.}

\item{obs.nodes}{Observation node depths. Numeric vector. Same length units as the project. Leave NULL if none wanted.}
}
\description{
Observation nodes (count and depths of each) are the last two rows of PROFILE.DAT
}
\examples{
\dontrun{
write.obs.nodes(project.path, obs.nodes = c(5, 10, 20, 30)) ## Writes at these depths
write.obs.nodes(project.path) ##
}

}
