% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_profile_data.R
\name{write.profile.data}
\alias{write.profile.data}
\title{Write profile data to profile.dat}
\usage{
write.profile.data(project.path, parameters)
}
\arguments{
\item{project.path}{Path of the HYDRUS1D project}

\item{parameters}{named list of single numbers or vectors to write to profile data section}
}
\value{
Write profile data to "profile.dat"
}
\description{
Write profile data to profile.dat
}
