% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rwu_stress.R
\name{write.rwu.stress}
\alias{write.rwu.stress}
\title{Write root water uptake variables}
\usage{
write.rwu.stress(
  project.path,
  model = 0,
  compensated.uptake = 1,
  P0 = -10,
  P2H = -200,
  P2L = -800,
  P3 = -8000,
  POptm = -25,
  r2H = 0.5,
  r2L = 0.1,
  ...
)
}
\arguments{
\item{project.path}{your project path, where SELECTOR.IN is saved.}

\item{model}{0 for Feddes, 1 for S-shape. Note that currently code only supports model = 0 and will stop if other value is given.}

\item{compensated.uptake}{Numeric, length = 1, value <= 1. Set equal to one for a noncompensated root water uptake and smaller than one for
compensated root water uptake. This function's default is the Hydrus default of 1, non-compensated. This parameter is labelled OmegaC in SELECTOR.IN.}

\item{P0, P2H, P2L, P3}{numeric, length = 1, value < 0. In same length units as project. Default values for "Pasture (Wesseling, 1981)" are given
here with cm units. These values are, respectively, the pressure heads below which roots begin to extract water, fall below the maximum
transpiration rate r2H, fall below the maximum transpiration rate r2L, and cease to extract water (aka wilting point).}

\item{POptm}{numeric, length = number of materials, value < 0. Value of the pressure head, h2, below which roots start to extract
water at the maximum possible rate. Values must be given in the same order as the material hydraulic properties.}

\item{r2H, r2L}{potential transpiration rates, in project units of length/time, associated with pressure heads between P0 and P2H, and between P2H and P2L, respectively.}
}
\description{
These are the pressur eheads at which roots increase, decrease, or cease water uptake. Default values oif this function are for Hydrus' default setting,
which is a Feddes model with parameters for "Pasture (Wesseling, 1981)". Alternate values can be found by looking in the 'Root Water Uptake -
Water Stress Reduction' menu of the Hydrus GUI. This function is currently not coded to process an S-shaped model.
}
\author{
Trevor Baker <tbaker@iegconsulting.com>
}
