% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_simulation_settings.R
\name{write.sim.settings}
\alias{write.sim.settings}
\title{Write simulation settings to SELECTOR.IN (Block B)}
\usage{
write.sim.settings(
  project.path,
  MaxIt = 100,
  TolTh = 0.001,
  TolH = 1,
  TopInf = TRUE,
  BotInf = FALSE,
  WLayer = TRUE,
  KodTop = -1,
  KodBot = -1,
  InitCond = FALSE,
  qGWLF = FALSE,
  FreeD = TRUE,
  SeepF = FALSE,
  DrainF = FALSE,
  hSeep = 0
)
}
\arguments{
\item{project.path}{path to project, where SELECTOR.IN file exists already (e.g. made by create.h1d.project() )}

\item{MaxIt}{Integer, length = 1. Maximum number of iterations allowed during any time step}

\item{TolTh}{Absolute water content tolerance for nodes in the unsaturated part of the flow region. Hydrus' recommended value is 0.0001, Default
here is 0.001 to allow quicker simulations. TolTh is the maximum desired absolute change in the value of the water content, theta, between two
successive iterations during a particular time step.}

\item{TolH}{Absolute pressure head tolerance for nodes in the saturated part of the flow region. Hydrus' recommended value is 0.1 cm, and the
default given here is 1 to allow quicker solutions (beware that the value given must be in the project's lenght units and will not automatically
be in cm). TolH represents the maximum desired absolute change in the value of the pressure head, h, between two successive iterations during
a particular time step.}

\item{TopInf, BotInf}{logical, length = 1. Default TopInf=TRUE, BotInf=FALSE. TRUE = time-dependent boundary condition is to be imposed at
the top/bottom of the profile, for which data are supplied via input file ATMOSPH.IN (created by write_atmosph_in function).}

\item{WLayer}{logical, length = 1. Default T. TRUE = water can accumulate at the surface with zero surface runoff.}

\item{KodTop, KodBot}{integer, length = 1. Codes specifying type of boundary condition (BC) for water flow at the surface and bottom of profile,
respectively. Code number is positive for Dirichlet BC (i.e., boundary is at prescribed pressure head) and negative for Neumann BC (i.e.,
prescribed volumetric flux entering/leaving. e.g., precip).  Set KodTop to 0 when a prescribed BC can change from
Dirichlet BC to Neumann BC and vice versa. \cr
KodTop/KodBot values: +1 = constant surface/bottom head; -1 = constant surface (incl. atmospheric)/bottom flux; +3 = variable surface/bottom head;
-3 = variable surface/bottom flux; -4 variable atmospheric BC (KodTop only).}

\item{InitCond}{llogical, length = 1. Default F. Unsure. Might be IInitW from manual: TRUE if the initial condition is given in terms of theta,
FALSE if given in terms of the pressure head.}

\item{qGWLF}{logical, length = 1. Default F. TRUE = discharge-groundwater level relationship q(GWL) is applied as bottom boundary condition}

\item{FreeD}{logical, length = 1. Default T.  TRUE = free drainage is to be considered as bottom boundary condition, i.e. unity gradient at bottom of
profile, meaning bottom layer is in matric equilibrium with the underlying material.}

\item{SeepF}{logical, length = 1. Default F. TRUE = seepage face is to be considered as the bottom boundary condition.}

\item{DrainF}{logical, length = 1. TRUE = a drain is to be simulated by means of a boundary condition}

\item{hSeep}{numeric, length = 1. Default = 0. Pressure head (i.e., 0) that initiates flow over the seepage face bottom boundary.}
}
\description{
Write settings controlling the number of iterations (MaxIt) and solver tolerances (TolTh, TolH), as well as defining simulation settings:
TopInf, WLayer, KodTop, InitCond, BotInf, qGWLF, FreeD, SeepF, KodBot, DrainF, hSeep.
}
