% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_nod_inf_out.R
\name{compile.nod_inf.out}
\alias{compile.nod_inf.out}
\title{Compile NOD_INF.OUT data from all simulations}
\usage{
compile.nod_inf.out(outputs.path, output = NULL, num.of.sims, ...)
}
\arguments{
\item{...}{}
}
\description{
Compile NOD_INF.OUT data from all simulations
}

