% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_H1D_project.R
\name{create.H1D.project}
\alias{create.H1D.project}
\title{Creates a new hydrus project folder with "HYDRUS1D.DAT" and "DISCRIPT.TXT" files}
\usage{
create.H1D.project(project.name, parent.dir, discription = NULL,
  TimeUnit = "days", SpaceUnit = "cm", PrintTimes = 1,
  processes = c(WaterFlow = T, RootWaterUptake = T), geometry, initial.cond,
  ...)
}
\arguments{
\item{parent.dir}{Path to the project folder}

\item{TimeUnit}{Simulation time unit information (default = days)}

\item{SpaceUnit}{Vertical spatial unit (decault = cm)}

\item{PrintTimes}{Time levels at which the outputs should be printed}

\item{processes}{Main processes (e.g., WaterFlow, SoluteTransport etc) to be simualted}

\item{geometry}{Profile geometry info (Depth, # of nodes and # of obs nodes)}
}
\description{
@param project.name Name of the project
}
\examples{

create.H1D.project(project.name = "testproj2", parent.dir = parent_dir, discription = NULL,
SpaceUnit = "cm", TimeUnit = "days", PrintTimes = 1,
processes = c(WaterFlow = T, SoluteTransport = F, RootWaterUptake = F,
RootGrowth = F, Unsatchem = F, HP1 = F, EquillibriumAdsorption = F,
initial.cond = c(NumberOfSolutes = 0, InitialCondition = 0),
geometry = c(ProfileDepth = 200, NumberOfNodes = 4,
         ObservationNodes = 0, MaterialNumber = 1, SubregionNumber = 1))
}
