% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_soil_profile.R
\name{create.soil.profile}
\alias{create.soil.profile}
\title{Creates a new PROFILE.DAT file using a template included in the package.}
\usage{
create.soil.profile(
  project.path,
  out.file = "PROFILE.DAT",
  profile.depth,
  dz = 1,
  Temp = 20,
  Conc = 0,
  obs.nodes = NULL
)
}
\arguments{
\item{project.path}{Location of the H1D project in the directory}

\item{out.file}{Default is always 'PROFILE.OUT'}

\item{profile.depth}{profile.depth}

\item{dz}{Descretization step of the profile. (default: 1)}

\item{Temp}{Temperate input default is 20 degree C}

\item{Conc}{Concentration, zero in initial profile (default: 0)}

\item{obs.nodes}{Vector of observation points in the profile (default: NULL)}
}
\value{
creates a new PROFILE.DAT
}
\description{
Creates a new PROFILE.DAT file using a template included in the package.
}
