% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_dir.R
\name{dir.copy}
\alias{dir.copy}
\title{Copy contents of a folder to another}
\usage{
dir.copy(src.dir, dest.dir, overwrite = TRUE, ask = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Copy contents of a folder to another
}
