% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_hourly.R
\name{et.hourly}
\alias{et.hourly}
\title{Generates hourly ET inputs from provided daily value.}
\usage{
et.hourly(Et.Daily, hours = 6:20)
}
\arguments{
\item{hours}{}
}
\description{
Generates hourly ET inputs from provided daily value.
}
