% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{genSelector}
\alias{genSelector}
\title{Modify a SELECTOR.IN template file with data for some of the records in the different blocks}
\usage{
genSelector(b17 = NULL, TPrint = NULL, chpar = NULL, F.Ks = NULL,
  F.SnkL1p = NULL, F11 = NULL, DifW = 0, DifG = 0, ...,
  template = "SELECTOR.IN", txt = readLines(template),
  blocks = getBlocks(txt), origNMat = as.integer(getValue(txt, "NMat")))
}
\arguments{
\item{b17}{A data frame with NMat rows, for Block B record 17.}

\item{TPrint}{a vector of times to print the results. For Block C, record 9}

\item{chpar}{Data frame giving solute transport parameters, 1 row per solute Block F, record 7}

\item{F.Ks}{a vector giving the Ks columns for Block F, record 11.}

\item{F.SnkL1p}{a vector giving the SnkL1' columns for Block F, record 11}

\item{F11}{a data frame providing all of the columns for each solute. Block F, record 11
F11 should have NMat * No.Solutes rows.}
}
\item{\dots}{\code{name=value} pairs of parameters to set. These
  are passed to \code{replaceValues}. 
}
\item{template}{the path to the template SELECTOR.IN file}
\item{txt}{the text of the template file}

\description{
This does not check the conditional logic based on the
values of parameters in the file such as Hysterisis, iModel, etc. which
change the structure of what is expected in the inputs.
This only allows the caller to specify a few records in the selector file.
However, the tools \code{replaceValues}, \code{replaceColumn}, \code{replaceData}
allow one to write code reasonably easily to handle other records.
}

\value{
a character vector containing the updated contents of the input file.
You need to write this to a file.  This differs from the original functions
in hydrusR.
Note that if you are using RHydrus, one can name the files differently from SELECTOR.IN, etc.
and specify the input files to use directly. This means that the names of the input files
can indicate their purpose/characteristics.
}

