% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval.R
\name{model.eval}
\alias{model.eval}
\title{Calculate RMSE and Nash-Sutcliffe CEFF of the simulation}
\usage{
model.eval(data, var.name = "", print.stats = TRUE, plot = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Calculate RMSE and Nash-Sutcliffe CEFF of the simulation
}
