% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_H1D_climate.R
\name{prep.H1D.climate}
\alias{prep.H1D.climate}
\title{Prepare climate data for Hydrus-1D}
\usage{
prep.H1D.climate(
  project.path,
  TimeUnit = "hours",
  SpaceUnit = "cm",
  endTime = NULL,
  time_values = NULL,
  PET_mmd = 4,
  PET_values = NULL,
  trans.pc = 1,
  Prec_mmd = 0.5,
  Prec_values = NULL,
  diurnal = TRUE,
  diurnal.hours = c(6, 20),
  save.df = FALSE
)
}
\arguments{
\item{project.path}{Hydrus project path. Must have SELECTOR.IN already, which is made by create.H1D.project().}

\item{TimeUnit}{character, length 1. your project's time unit (default = "hours"). permitted: "seconds", "minutes", "hours", "days", "years".}

\item{SpaceUnit}{character, length 1. your project's space (length) unit (default = "cm"). permitted: "mm", "cm", "m".}

\item{endTime}{numeric, length = 1. what is the ending time of your simulation? in project TimeUnits.}

\item{time_values}{numeric, any length. Last row must be for the maximum time (endTime) of your simulation. Given in your project's TimeUnit. Two
formats of climate data inputs are allowed to this function: time_values is given with one or both of PPT_values and Prec_values; or time_values
can be omitted (left as NULL) and simple daily rates of PET_mmd and/or Prec_mmd can be given.}

\item{PET_mmd}{numeric, length 1. PET rate in mm/d. always in mm/d, no matter your project's units. This can be omitted if 'PET_values' is given.}

\item{PET_values}{numeric vector, any length. Potential evapotranspiration values at each interval listed in 'time_values'. Length must match the
length of 'time_values'. Unlike PET_mmd, which is alwys in mm/d, this vector is in project units, SpaceUnit per TimeUnit.}

\item{trans.pc}{numeric, length = 1, between 0 and 1. If PET is a column in atmos.df, what percent of PET should be assigned to transpiration
(rRoot)? The remainder will be assigned to evaporation (rSoil). Default = 1, i.e. all transpiration. Not required if 'PET' column is not given in
atmos.df.}

\item{Prec_mmd}{numeric, length 1. Precipitation rate in mm/d. Always in mm/d, no matter your project's units. This can be omitted if 'Prec_values' is given.}

\item{Prec_values}{numeric vector, any length. Precipitation values at each interval listed in 'time_values'. Length must match the length of
'time_values'. Unlike Prec_mmd, which is alwys in mm/d, this vector is in project units, SpaceUnit per TimeUnit.}

\item{diurnal}{logical, length 1. If your TimeUnit is less than "days": TRUE - PET is partitioned into diurnal patterns by pet.hourly(); FALSE -
distributed evenly across all hours (FALSE). Ignored if TimeUnit = "days" or "years"}

\item{diurnal.hours}{integer, length 2. Default c(6,20). Giving ~sunrise and ~sunset times in 24-hour format. The first hour with non-zero PET in
the morning, and the first hour with zero PET at night. Ignored if diurnal = FALSE or TimeUnit = "days" or "years".}

\item{save.df}{logical, length 1. Default FALSE. should this climate data be saved as 'df_clim.csv' to the project.path? The reason to do this
would be for long simulations that need to utilize the looping function run.H1D.simulation. With code edits there, it could read the csv file to
select climate data for the next loop iteration.}
}
\description{
Prepare climate data for Hydrus-1D. Returns a dataframe and optionally writes it to file in your project directory.
}
