% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_node_inf.R
\name{read.nod_inf}
\alias{read.nod_inf}
\title{READ outputs of NOD_INF.OUT}
\usage{
read.nod_inf(project.path, out.file = "Nod_Inf.out", output = NULL,
  warn = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
READ outputs of NOD_INF.OUT
}
