% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obs_node.R
\name{read.obs_node}
\alias{read.obs_node}
\title{Read observation point outputs from}
\usage{
read.obs_node(project.path, out.file = "Obs_Node.out", obs.output = NULL,
  obs.nodes.all, ...)
}
\arguments{
\item{...}{}
}
\description{
Read observation point outputs from
}
