% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tlevel_out.R
\name{read.T_level.out}
\alias{read.T_level.out}
\title{READ outputs of TLEVEL.OUT}
\usage{
read.T_level.out(simulation.path, output = NULL, warn = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
READ outputs of TLEVEL.OUT
}

