% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_H1D_simulation.R
\name{run.H1D.simulation}
\alias{run.H1D.simulation}
\title{This is the main simulation function}
\usage{
run.H1D.simulation(project.path, hydrus.path = NULL, profile.depth, beginT,
  endT, deltaT, bot.bc.type, bot.bc.value, const.bot.bc, soil.para, atm.bc.data,
  ini.wt, TimeUnit = "days", rdepth, obs.nodes, show.output = TRUE, ...)
}
\arguments{
\item{endT}{end time}

\item{deltaT}{time step}

\item{bot.bc.type}{head or flux type}

\item{bot.bc.value}{value of the bc}

\item{const.bot.bc}{Logical, to set if bottom BC is constant}

\item{soil.para}{Hydraulic parameters of soil (van Genuchten)}

\item{atm.bc.data}{data frame containing atmonspheric boundary conditions (time variable BC)}

\item{ini.wt}{Initial water table depth}

\item{rdepth}{rooting depth}

\item{obs.nodes}{Observation node points (vector)}

\item{show.output}{Logical, whether the shell output of HYDRUS1D run should be displayed on R console, default = F}
}
\description{
This is the main simulation function
}
