% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_atmosph_in.R
\name{write.atmosph.in}
\alias{write.atmosph.in}
\title{Write atmospheric boundary condition inputs}
\usage{
write.atmosph.in(
  project.path,
  maxAL,
  deltaT,
  atm.bc.data,
  hCritS = 0,
  input.pet = T,
  LAI = 0.39
)
}
\arguments{
\item{project.path}{project.path}

\item{maxAL}{maxAL}

\item{deltaT}{deltaT}

\item{atm.bc.data}{atm.bc.data}

\item{hCritS}{default: 0}

\item{input.pet}{default: T}

\item{LAI}{default: 0.39}
}
\value{
write ATMOSPH.IN
}
\description{
Write atmospheric boundary condition inputs
}
