% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_atmosph_in.R
\name{write.atmosph.in}
\alias{write.atmosph.in}
\title{Write atmospheric boundary condition inputs}
\usage{
write.atmosph.in(file.atmosph.in = "ATMOSPH.IN", MaxAL, tstep, atm.bc.data,
  sim_ind = 1, ...)
}
\arguments{
\item{...}{}
}
\description{
Write atmospheric boundary condition inputs
}

