% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_atmosph_in.R
\name{write.atmosph.in}
\alias{write.atmosph.in}
\title{Write atmospheric boundary condition inputs}
\usage{
write.atmosph.in(project.path, maxAL, deltaT, atm.bc.data, hCritS = 0,
  input.pet = T, LAI = 0.39, ...)
}
\arguments{
\item{...}{}
}
\description{
Write atmospheric boundary condition inputs
}
