% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bottom_bc.R
\name{write.bottom.bc}
\alias{write.bottom.bc}
\title{Write bottom Boundary Condition}
\usage{
write.bottom.bc(project.path, constant.bc, bc.type, bc.value)
}
\arguments{
\item{project.path}{Path of the HYDRUS1D project}

\item{constant.bc}{constant.bc}

\item{bc.type}{bc.type}

\item{bc.value}{bc.value}
}
\value{
Write bottom Boundary Condition to "SELECTOR.IN"
}
\description{
Write bottom Boundary Condition
}
