% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bottom_bc.R
\name{write.bottom.bc}
\alias{write.bottom.bc}
\title{Write bottom BC}
\usage{
write.bottom.bc(constant.bc, bc.type, bc.value, project.path, ...)
}
\arguments{
\item{...}{}
}
\description{
Write bottom BC
}

