% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hydraulic_para.R
\name{write.hydraulic.para}
\alias{write.hydraulic.para}
\title{Write soil hydraulic parameters to selector.in}
\usage{
write.hydraulic.para(project.path, model = 0, hysteresis = 0, para)
}
\arguments{
\item{project.path}{Path of the HYDRUS1D project}

\item{model}{Integer (default: 0) defining which model should be used}

\item{hysteresis}{Integer (default: 0) indicating whether hysteresis is considered}

\item{para}{Named list of parameters. Each parameter can be a single number or a vector if there is more than one material}
}
\value{
Write soil hydraulic parameters to "selector.in"
}
\description{
Write soil hydraulic parameters to selector.in
}
\details{
Available soil moisture retention models in the hydrus1D:
van Genuchten (VG)
6 parameter van Genuchten (VGM),
Brooks Corey (BC)
Kosugi model (KOS)
Dual porosity model (DUR)
Dual porosity model 2 (DUR2)
Dual porosity model 3 (DUR3)
Details on the models on hydrus help manual and elsewhere in the literature
}
