% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_initial_condition.R
\name{write.ini.cond}
\alias{write.ini.cond}
\title{Write initial conditions in profile.dat}
\usage{
write.ini.cond(file.profile.dat, pr.vec = NULL, wt, soil.depth, n.nodes, ...)
}
\arguments{
\item{...}{}
}
\description{
Write initial conditions in profile.dat
}

