% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_initial_condition.R
\name{write.ini.cond}
\alias{write.ini.cond}
\title{Write root distribution in profile.dat}
\usage{
write.ini.cond(project.path, pr.vec = NULL, wt.depth, ...)
}
\arguments{
\item{project.path}{Path of HYDRUS1D project}

\item{pr.vec}{A vector of pressure head values (length = # of total nodes in the profile)}

\item{wt.depth}{Depth of water table (to assign hydrostatic initial condition)}

\item{...}{}

\item{profile.depth}{Depth of soil profile}
}
\value{

}
\description{
Write root distribution in profile.dat
}
