% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_obs_nodes.R
\name{write.obs.nodes}
\alias{write.obs.nodes}
\title{Write observation nodes in profile.dat}
\usage{
write.obs.nodes(project.path, obs.nodes = NULL, ...)
}
\arguments{
\item{project.path}{path to the hydrus project}

\item{obs.nodes}{Observation nodes}
}
\description{
Write observation nodes in profile.dat
}
\examples{
\dontrun{
write.obs.nodes(project.path, obs.nodes = c(5, 10, 20, 30)) ## Writes at these depths
write.obs.nodes(project.path) ##
}

}
