% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_observation_nodes.R
\name{write.obs.nodes}
\alias{write.obs.nodes}
\title{Write observation nodes in profile.dat}
\usage{
write.obs.nodes(project.path, Z, dz, obs.nodes)
}
\arguments{
\item{obs.nodes}{}
}
\description{
Write observation nodes in profile.dat
}

