% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_print_times.R
\name{write.print.times}
\alias{write.print.times}
\title{Write print times in SELECTOR.IN}
\usage{
write.print.times(project.path, tmin, tmax, tstep, TimeUnit, ...)
}
\arguments{
\item{project.path}{Location of the H1D project in the directory}

\item{tmin}{Begining print time (e.g., 1)}

\item{tmax}{End print time (e.g. 240 hours)}

\item{tstep}{Time step to print out put at}

\item{...}{}
}
\description{
Write print times in SELECTOR.IN
}
