% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_print_times.R
\name{write.print.times}
\alias{write.print.times}
\title{Write print times in selector.in}
\usage{
write.print.times(input.file, tmax, tstep, ...)
}
\arguments{
\item{...}{}
}
\description{
Write print times in selector.in
}

