% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_print_times.R
\name{write.print.times}
\alias{write.print.times}
\title{Update print times, MPL, tInit, and tMax in SELECTOR.IN}
\usage{
write.print.times(project.path, print.at)
}
\arguments{
\item{project.path}{Location of the H1D project in the directory}

\item{print.at}{numeric vector, any length. The times to print results at, in this project's time units.}
}
\description{
This function has two uses. To create a properly formatted TPrint block for SELECTOR.IN containing all times that values will be printed (e.g., to
"Nod_Inf.OUT"), and, for use in hydrusR's looping of long simulations, to replace the time range and MPL values for each successive loop's input file.
}
