% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_root_dist.R
\name{write.root.dist}
\alias{write.root.dist}
\title{Write root distribution in profile.dat}
\usage{
write.root.dist(project.path, rdepth, rBeta = 0.962)
}
\arguments{
\item{project.path}{project.path}

\item{rdepth}{root depth}

\item{rBeta}{default: 0.962}
}
\value{
Write root distribution in "PROFILE.DAT"
}
\description{
Write root distribution in profile.dat
}
