% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_root_dist.R
\name{write.root.dist}
\alias{write.root.dist}
\title{Write root distribution in profile.dat}
\usage{
write.root.dist(file.profile.dat, rdepth, rbeta = 0.962, n.nodes, ...)
}
\arguments{
\item{...}{}
}
\description{
Write root distribution in profile.dat
}

