% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_time_settings.R
\name{write.time.settings}
\alias{write.time.settings}
\title{Write time settings}
\usage{
write.time.settings(
  project.path,
  time.range = c(0, 2400),
  dt = 0.001,
  dtMin = 1e-06,
  dtMax = 1,
  DMul = c(0.7, 1.3),
  ItRange = c(3, 7),
  print.step
)
}
\arguments{
\item{project.path}{your project path, where SELECTOR.IN is saved.}

\item{time.range}{numeric, length = 1. in your project's time units. the start and end times of this simulation.}

\item{dt}{numeric, length = 1. initial time step, in your project's time units.}

\item{dtMin, dtMax}{numeric, length = 1. minimum and maximum  allowed time step, in your project's time units.}

\item{DMul}{numeric, length = 2. range of time step multiplication factors. Default c(0.7,1.3).}

\item{ItRange}{integer, length = 2. optimal iteration range. Sets parameters ItMin and ItMax.}

\item{print.step}{integer, length = 1. the interval at which results will be printed (e.g. to "Nod_Inf.OUT"). in your project's time units.}
}
\description{
These are the settings from Time Information-Time Discretization and Water Flow-Iteration Criteria in the Hydrus GUI. The time settings that control
iterations, time steps, run times, etc. These get written to Block C of SELECTOR.IN.
}
\author{
Trevor Baker <tbaker@iegconsulting.com>
}
